仓库批量设置修改

+++ middle-web/src/main/java/com/pousheng/middle/web/shop/convert/WarehouseRuleItemConverter.java
@@ -87,7 +87,7 @@ public class WarehouseRuleItemConverter {
         List<Long> list = Lists.newArrayList();
         for (WarehouseRuleItem obj1 : exist) {
             for (WarehouseRuleItem obj2 : addList) {
-                if (obj1.getWarehouseId() == obj2.getWarehouseId()) {
+                if (obj1.getWarehouseId().longValue() == obj2.getWarehouseId().longValue()) {
                     list.add(obj1.getWarehouseId());
                     continue;
                 }
